border on the way WIP

+++ projects/fender/src/Systems/SFML/Grid.cpp
@@ -24,15 +24,15 @@ namespace fender::systems::SFMLSystems {
                 sf::Vertex(sf::Vector2f(window.getSize().x, 0))
         };
         auto &camPos = cam.getEntity().get<components::Transform>();
-        intptr_t x{-(int)camPos.position.x % world->unit};
-        intptr_t y{-(int)camPos.position.y % world->unit};
-        while (x < window.getSize().x) {
+        int x{-(int)camPos.position.x % world->unit};
+        int y{-(int)camPos.position.y % world->unit};
+        while (x < (int)window.getSize().x) {
             vertical[0].position.x = x;
             vertical[1].position.x = x;
             window.draw(vertical, 2, sf::Lines);
             x += world->unit;
         }
-        while (y < window.getSize().y)
+        while (y < (int)window.getSize().y)
         {
             horizontal[0].position.y  = y;
             horizontal[1].position.y = y;
